+++ app/src/main/java/com/ruuvi/station/units/domain/UnitsConverter.kt
@@ -469,13 +469,13 @@ class UnitsConverter (
             unitType = NOX.NoxIndex
         )
 
-    fun getLuminosityEnvironmentValue(luminosity: Int) =
+    fun getLuminosityEnvironmentValue(luminosity: Double) =
         EnvironmentValue(
-            original = luminosity.toDouble(),
-            value = luminosity.toDouble(),
+            original = luminosity,
+            value = luminosity,
             accuracy = Accuracy.Accuracy1,
-            valueWithUnit = context.getString(Accuracy.Accuracy0.nameTemplateId, luminosity.toDouble(), context.getString(Luminosity.Lux.unit)),
-            valueWithoutUnit = context.getString(Accuracy.Accuracy0.nameTemplateId, luminosity.toDouble(), "").trim(),
+            valueWithUnit = context.getString(Accuracy.Accuracy0.nameTemplateId, luminosity, context.getString(Luminosity.Lux.unit)),
+            valueWithoutUnit = context.getString(Accuracy.Accuracy0.nameTemplateId, luminosity, "").trim(),
             unitString = context.getString(Luminosity.Lux.unit),
             unitType = Luminosity.Lux
         )
